+++ src/domain/entity/room.go
@@ -8,25 +8,25 @@ import (
 type Room struct {
 	Id       gocql.UUID `cql:"id" json:"id" faker:"-"`
 	RoomId   int64      `cql:"room_id" json:"room_id"`
-	Status   int32      `cql:"status" json:"status"`
+	Status   int64      `cql:"status" json:"status"`
 	UsersId  []int64    `cql:"users_id" json:"users_id"`
 	Messages []Message  `cql:"messages" json:"messages"`
 	MetaData []MetaData `cql:"meta_data" json:"meta_data"`
-	Rating   int32      `cql:"rating" json:"rating"`
+	Rating   int64      `cql:"rating" json:"rating"`
 }
 
 func (e *Room) GetId() gocql.UUID                     { return e.Id }
 func (e *Room) SetId(id gocql.UUID) *Room             { e.Id = id; return e }
 func (e *Room) GetRoomId() int64                      { return e.RoomId }
 func (e *Room) SetRoomId(roomId int64) *Room          { e.RoomId = roomId; return e }
-func (e *Room) GetStatus() int32                      { return e.Status }
-func (e *Room) SetStatus(status int32) *Room          { e.Status = status; return e }
+func (e *Room) GetStatus() int64                      { return e.Status }
+func (e *Room) SetStatus(status int64) *Room          { e.Status = status; return e }
 func (e *Room) GetUsersId() []int64                   { return e.UsersId }
 func (e *Room) SetUsersId(usersId []int64) *Room      { e.UsersId = usersId; return e }
 func (e *Room) GetMessages() []Message                { return e.Messages }
 func (e *Room) SetMessages(messages []Message) *Room  { e.Messages = messages; return e }
-func (e *Room) GetRating() int32                      { return e.Rating }
-func (e *Room) SetRating(rating int32) *Room          { e.Rating = rating; return e }
+func (e *Room) GetRating() int64                      { return e.Rating }
+func (e *Room) SetRating(rating int64) *Room          { e.Rating = rating; return e }
 func (e *Room) Factory() error                        { return faker.FakeData(e) }
 func (e *Room) GetMetaData() []MetaData               { return e.MetaData }
 func (e *Room) SetMetaData(metaData []MetaData) *Room { e.MetaData = metaData; return e }
