+++ src/domain/http/request/addMessage.go
@@ -7,7 +7,7 @@ type AddMessage struct {
 	RoomId       int64  `json:"room_id" validate:"required"`
 	UserId       int64  `json:"user_id" validate:"required"`
 	Content      string `json:"content" validate:"required"`
-	Kind         int32  `json:"kind" validate:"required"`
+	Kind         int64  `json:"kind" validate:"required"`
 }
 
 func (r *AddMessage) Validate() error {
