+++ src/domain/http/request/addMetaData.go
@@ -6,7 +6,7 @@ type AddMetadata struct {
 	CustomerName string `json:"customer_name" validate:"required"`
 	RoomId       int64  `json:"room_id" validate:"required"`
 	Key          string `json:"key" validate:"required"`
-	Kind         int32  `json:"kind" validate:"required"`
+	Kind         int64  `json:"kind" validate:"required"`
 	Value        string `json:"value" validate:"required"`
 }
